@@ -19,34 +19,35 @@ public class SapAmqpConnectorConfiguration {
     private String uri;
 
 
-    @Parameter
-    @DisplayName("Token Endpoint URL")
-    @Example("https://your-subdomain.authentication.eu20.hana.ondemand.com/oauth/token")
-    private String tokenUrl;
-
+//    @Parameter
+//    @DisplayName("Token Endpoint URL")
+//    @Example("https://your-subdomain.authentication.eu20.hana.ondemand.com/oauth/token")
+//    private String tokenUrl;
+//
     @Parameter
     @DisplayName("Client ID")
     private String clientId;
 
-    @Parameter
-    @DisplayName("Client Secret")
-    @Password
-    private String clientSecret;
-
+//    @Parameter
+//    @DisplayName("Client Secret")
+//    @Password
+//    private String clientSecret;
+//
     // Getters
     public String getUri() { 
         return uri; 
     }
-    
-    public String getTokenUrl() { 
-        return tokenUrl; 
-    }
+//    
+//    public String getTokenUrl() { 
+//        return tokenUrl; 
+//    }
     
     public String getClientId() { 
         return clientId; 
     }
     
-    public String getClientSecret() { 
-        return clientSecret; 
-    }
+//    public String getClientSecret() { 
+//        return clientSecret; 
+//    }
+//    
 }
